refactor: move PinsecTester
